Basic debugging is working. DW_OP_addr is working, good for global variables.

+++ deploy/debugger/arduino/soft_debugger.cpp
@@ -33,7 +33,7 @@ typedef struct
 typedef struct 
 {
     DbgHeader Header;
-    uint16_t  Address;
+    uint32_t  Address;
     uint8_t   Size;
 } DbgTraceQueryPacket;
 
@@ -45,7 +45,7 @@ typedef struct
 
 
 
-static void DbgSendTrace(uint16_t address, uint8_t size);
+static void DbgSendTrace(uint32_t address, uint8_t size);
 
 
 void DbgConnect()
@@ -75,7 +75,7 @@ static void DbgLoop()
         {
             uint16_t address;
             uint8_t size;
-            Serial.readBytes((char*)&address, sizeof(uint16_t));
+            Serial.readBytes((char*)&address, sizeof(uint32_t));
             Serial.readBytes((char*)&size, sizeof(uint8_t));
             
             DbgSendTrace(address, size);
@@ -101,7 +101,7 @@ void DbgBreak(uint8_t breakpointNo)
     DbgLoop();
 }
 
-static void DbgSendTrace(uint16_t address, uint8_t size)
+static void DbgSendTrace(uint32_t address, uint8_t size)
 {
     DbgTraceAnswerPacket p;
